@@ -4,14 +4,14 @@ description: A starting point for Dart libraries or applications.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  dartz: 0.9.2
-  riverpod: 0.12.2
+  dartz: ^0.10.0-nullsafety.1
+  riverpod: ^0.14.0+3
   clima_core:
     path: ../clima_core
 
 dev_dependencies:
-  test: ^1.14.4
-  lint: ^1.4.0-dev.d210.1
+  test: ^1.17.3
+  lint: ^1.5.3
