@@ -12,9 +12,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  auto_size_text: ^2.1.0
-  charts_flutter: ^0.6.0
-  flutter_swiper: ^1.1.6
+  auto_size_text: 2.1.0
+  charts_flutter: 0.10.0
+  flutter_swiper: 1.1.6
 
   clima_core:
     path: ../clima_core
@@ -23,23 +23,23 @@ dependencies:
   clima_data:
     path: ../clima_data
 
-  connectivity: ^2.0.1
-  dartz: 0.9.2
-  equatable: 1.2.5
-  flutter_hooks: 0.15.0
-  flutter_riverpod: 0.12.2
-  flutter_spinkit: ^4.1.2+1
-  flutter_weather_icons: ^1.0.2
-  hooks_riverpod: 0.12.2
-  http: ^0.12.0+2
-  material_floating_search_bar: ^0.2.6
-  package_info: ^0.4.3+4
-  riverpod: ^0.12.1
-  shared_preferences: ^0.5.12+4
-  font_awesome_flutter: ^8.12.0
+  connectivity: 3.0.3
+  dartz: 0.10.0-nullsafety.1
+  equatable: 2.0.0
+  flutter_hooks: 0.16.0
+  flutter_riverpod: 0.14.0+3
+  flutter_spinkit: 5.0.0
+  flutter_weather_icons: 1.0.2
+  hooks_riverpod: 0.14.0+3
+  http: 0.13.2
+  material_floating_search_bar: 0.3.4
+  package_info: 2.0.0
+  riverpod: 0.14.0+3
+  shared_preferences: 2.0.5
+  font_awesome_flutter: 9.0.0
 
 dev_dependencies:
-  lint: ^1.4.0-dev.d210.1
+  lint: 1.5.3
   flutter_test:
     sdk: flutter
 
