@@ -1,11 +1,10 @@
 import 'package:clima_domain/entities/weather.dart';
-import 'package:meta/meta.dart';
 import 'package:equatable/equatable.dart';
 
 class Forecasts extends Equatable {
   const Forecasts({
-    @required this.forecasts,
-    @required this.cityName,
+    required this.forecasts,
+    required this.cityName,
   });
 
   final List<Weather> forecasts;
@@ -13,5 +12,5 @@ class Forecasts extends Equatable {
   final String cityName;
 
   @override
-  List<Object> get props => [forecasts, cityName];
+  List<Object?> get props => [forecasts, cityName];
 }
