@@ -5,7 +5,6 @@ import 'package:clima_domain/entities/forecasts.dart';
 import 'package:clima_domain/repos/forecasts_repo.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:riverpod/riverpod.dart';
 
 class GetForecasts implements UseCase<Forecasts, GetForecastsParams> {
@@ -19,12 +18,12 @@ class GetForecasts implements UseCase<Forecasts, GetForecastsParams> {
 }
 
 class GetForecastsParams extends Equatable {
-  const GetForecastsParams({@required this.city});
+  const GetForecastsParams({required this.city});
 
   final City city;
 
   @override
-  List<Object> get props => [city];
+  List<Object?> get props => [city];
 }
 
 final getForecastsProvider =
