@@ -13,5 +13,5 @@ class NoParams extends Equatable {
   const NoParams();
 
   @override
-  List<Object> get props => [];
+  List<Object?> get props => const [];
 }
