@@ -7,7 +7,7 @@ import 'package:clima_ui/utilities/hooks.dart';
 import 'package:clima_ui/utilities/reusable_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -20,9 +20,10 @@ class LoadingScreen extends HookWidget {
   Widget build(BuildContext context) {
     final _themeState = context.read(themeStateNotifier);
     final scaffoldKey = useGlobalKey<ScaffoldState>();
-    final weatherStateNotifier = useProvider(weatherStateNotifierProvider);
+    final weatherStateNotifier =
+        useProvider(weatherStateNotifierProvider.notifier);
     final forecastsStateNotifier =
-        useProvider(f.forecastsStateNotifierProvider);
+        useProvider(f.forecastsStateNotifierProvider.notifier);
 
     useEffect(
       () {
