@@ -7,7 +7,7 @@ import 'package:clima_domain/use_cases/get_forecasts.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
-import 'package:riverpod/all.dart';
+import 'package:riverpod/riverpod.dart';
 import 'package:state_notifier/state_notifier.dart';
 
 @sealed
@@ -91,7 +91,8 @@ class ForecastsStateNotifier extends StateNotifier<ForecastsState> {
 }
 
 final forecastsStateNotifierProvider =
-    StateNotifierProvider((ref) => ForecastsStateNotifier(
-          ref.watch(getForecastsProvider),
-          ref.watch(getCityProvider),
-        ));
+    StateNotifierProvider<ForecastsStateNotifier, ForecastsState>(
+        (ref) => ForecastsStateNotifier(
+              ref.watch(getForecastsProvider),
+              ref.watch(getCityProvider),
+            ));
