@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
-import 'package:riverpod/riverpod.dart';
 
 import 'current_conditions_widgets.dart';
 import 'forecast_widget.dart';
@@ -14,7 +13,7 @@ class WeatherWidget extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
-    final weather = useProvider(weatherStateNotifierProvider.state).weather;
+    final weather = useProvider(weatherStateNotifierProvider).weather;
 
     if (weather == null) {
       return const SizedBox.shrink();
